feat: verifica se existe o env para cria-lo

+++ application/app/Middlewares/GenerateKeysMiddleware.php
@@ -6,7 +6,7 @@
  * @author Vagner Cardoso <vagnercardosoweb@gmail.com>
  * @link https://github.com/vagnercardosoweb
  * @license http://www.opensource.org/licenses/mit-license.html MIT License
- * @copyright 01/03/2020 Vagner Cardoso
+ * @copyright 05/03/2020 Vagner Cardoso
  */
 
 namespace App\Middlewares;
@@ -39,11 +39,11 @@ class GenerateKeysMiddleware extends Middleware
         foreach (['APP_KEY', 'API_KEY', 'DEPLOY_KEY'] as $key) {
             $value = Env::get($key, null);
             $value = Helper::normalizeValueType($value);
+            $pathEnv = Path::app('/.env');
 
-            if (empty($value)) {
+            if (empty($value) && !file_exists($pathEnv)) {
                 $quote = preg_quote("={$value}", '/');
                 $random = Str::randomBytes(32);
-                $pathEnv = Path::app('/.env');
 
                 file_put_contents(
                     $pathEnv,
